Update patch set 8
Patch Set 8:

This is not building with the necessary change to the interface because the build processes pulls directly from upstream. I thought specifying Depends-On: If19f3dd454d8726b8b8d8ca4b105532219f0f3f0 would take care of that but apparently not.

Is there a way to handle this since the charm doesn't appears to track the version of interfaces that it depends on? In this case both If19f3dd454d8726b8b8d8ca4b105532219f0f3f0 and this change need to be made at the same time to avoid failures.

Patch-set: 8
Label: Verified=0