Update patch set 1
Patch Set 1:

There are two bugs here; one in the lock-file and one in charm-tools.  The lock-file for "interface:ceph-mds" has a null branch (which is odd), and charm-tools crashes on it.

I'll fix the lock-file.  I think there is also a bug in the branch tracking code in that it uses master which I don't think it should (it should use the commit hash).  I'll raise bugs for that too.

Patch-set: 1