Pin pip<20.3 and setuptools<50.0.0
osci runs the CI jobs on Bionic, this version of tox (2.5) will
automatically upgrade the packages at virtualenv creation (see
https://github.com/tox-dev/tox/issues/448 ) and any pinning won't have
effect.

The recently version of setuptools (58.0) dropped the support for
use_2to3=true which is needed to install blessings (an indirect
dependency of charm-tools).

This patch adds pip.sh which it's a wrapper of pip that will (re)install
the versions of pip and setuptools needed to run this job.

Change-Id: I9b0b8f953ff47816977b99fd01d74469024a04bb